Add separate file for all API interaction actions, to avoid importing axios more than once
