+++ src/actions/auth.js
@@ -1,9 +1,4 @@
-import axios from 'axios';
 import { AuthActionType } from '../reducers/authReducer';
-import { clearUserProfile, startFetchUserProfile } from './applicationState';
-import { startFetchPostings } from './postings';
-
-const API_URL = process.env.REACT_APP_API_URL;
 
 export const signIn = data => ({
   type: AuthActionType.SIGN_IN,
@@ -12,7 +7,7 @@ export const signIn = data => ({
   uid: data.uid,
 });
 
-const signingIn = () => ({
+export const signingIn = () => ({
   type: AuthActionType.SIGNING_IN,
 });
 
@@ -25,7 +20,7 @@ export const signInFailed = error => ({
   error,
 });
 
-const signingUp = () => ({
+export const signingUp = () => ({
   type: AuthActionType.SIGNING_UP,
 });
 
@@ -34,75 +29,14 @@ export const signUpFailed = error => ({
   error,
 });
 
-const clearLocalStorage = () => {
+export const clearLocalStorage = () => {
   localStorage.removeItem('client');
   localStorage.removeItem('access-token');
   localStorage.removeItem('uid');
 };
 
-const processSignInResponse = response => ({
+export const processSignInResponse = response => ({
   accessToken: response.headers['access-token'],
   client: response.headers.client,
   uid: response.headers.uid,
 });
-
-export const startSignIn = data => dispatch => {
-  dispatch(signingIn());
-  axios({
-    method: 'post',
-    url: `${API_URL}/auth/sign_in`,
-    data: {
-      email: data.email,
-      password: data.password,
-    },
-  }).then(response => {
-    if (response.status === 401) dispatch(signInFailed());
-    else {
-      dispatch(signIn(processSignInResponse(response)));
-      dispatch(startFetchUserProfile());
-      dispatch(startFetchPostings());
-    }
-  }).catch(error => {
-    if (error.response.status === 401) dispatch(signInFailed('Invalid login credentials. Please try again.'));
-  });
-};
-
-export const startSignUp = data => dispatch => {
-  dispatch(signingUp());
-  axios({
-    method: 'post',
-    url: `${API_URL}/auth`,
-    data: {
-      email: data.email,
-      password: data.password,
-      password_confirmation: data.password_confirmation,
-    },
-  }).then(response => {
-    dispatch(signIn(processSignInResponse(response)));
-    dispatch(startFetchUserProfile());
-    dispatch(startFetchPostings());
-  }).catch(error => {
-    dispatch(signUpFailed(error.response.data.errors.full_messages.join(', ')));
-  });
-};
-
-export const startSignOut = () => (dispatch, getState) => {
-  const { uid, client, accessToken } = getState().auth;
-  axios({
-    method: 'delete',
-    url: `${API_URL}/auth/sign_out`,
-    headers: {
-      uid,
-      client,
-      'access-token': accessToken,
-    },
-  }).then(() => {
-    clearLocalStorage();
-    dispatch(signOut());
-    dispatch(clearUserProfile());
-  }).catch(() => {
-    clearLocalStorage();
-    dispatch(clearUserProfile());
-    dispatch(signOut());
-  });
-};
